refactor: extract availability mixin from coordinator core
